Adding support for explosion mat files.
Explosions require special handling. It would be clumsy to create one mat file
per animation frame, so, instead, animated mat files only contain the name of
the group as in dat-2000.mat, dat-2001.mat, etc.

However, for some reason, each frame in an explosion animation is loaded
multiple times by the game, whenever a mission is loaded -- not when the texture
is used in-game. So I added a new boolean to mute the debug messages to avoid
spam.