Procedural explosions are now mixed with animated textures.
The DS2 core explosion is still fully procedural, but regular explosions are now
mixed with procedural explosions.

If AlphaBlendEnabled is set to 1, then procedural explosions are mixed with the
animated textures. If set to 0, only procedural explosions are shown. I need a
way to show the original textures as well.