Fix the explosion alpha to enable proper blending with procedural explosions.
So it turns out that this game makes the fully-transparent areas of the
explosion textures white. Thus when blending with the procedural explosions,
that causes problems. The fix I went with here is to pre-multiply the alpha with
the color, that way transparent areas become black and can be blended properly
using the Screen blending mode or the direct addition mode, etc.