Fixed Flashing Laser Lights + version bump to 1.1.5.1
This is the same fix that was applied to the master branch, but this time on the
stable release. backProjectMetric should use a UINT for the index, not a WORD
(otherwise the index will overflow when there's more than 64k indices on
display on the screen).

This should be the last release on this branch. Just a quick patch while we test
the master branch properly.