Hyperspace Tunnel speed fix, version bump to 1.1.5.2
The hyperspace tunnel will go slowly backwards after rendering one procedural
explosion. This happens because explosions modify the tunnel_speed setting and
leave it at -1. The fix is to store this value in a different global and restore
it when we're rendering the hyperspace tunnel.

Also, version bump to 1.1.5.2. No new development here, just fix really ugly
bug in the previous version.