Add the DirectBVH4 GPU builder to the list of builders.
It builds almost as fast as the FastLQBVH and performs a little better in some
cases. The Refit4() still needs to be checked as I had to run a re-fit operation
as a post-proc step which shouldn't be necessary.

(cherry picked from commit d81b35babaecc5e90a6597ac42d0c0d77fd2c1a5)