Prevent underflow when releasing ComPtr objects.
This prevents an underflow that happens in these situations:

* The game displays a 3D object in the Tech Room and then we close the game.
* The game runs a mission and then we exit the game.

This underflow may also happen in other situations, and it also happens in
Jeremy's Golden ddraw.

However, the downside of this "fix" is that now there appears to be a small
memory leak (it might be a false positive, since the IDXGIDebug1 interface is
always reported as a leak).

To see the underflow, undo this "fix", enable the DirectX Debug layer through
DXCPL:

https://github.com/morallo/xwa_ddraw_d3d11/wiki/Debugging-XWA#graphical-interface-dxcpl

and install the debug version of ddraw.

(cherry picked from commit f204d050048bb0ee5719332c128f13504012b44d)