Fix Normal Mapping issues when restarting a mission by pressing H.
g_meshToFGMap must be cleared in DeviceResources::OnSizeChanged(), instead of
FlightStart().

The reason for that is that pressing H to restart a mission will cause the
D3dRenderer class to be re-created; but D3dRendererOptNodeHook() will not
get called again and that's where g_meshToFGMap is populated. Hence, if we clear
g_meshToFGMap in D3dRenderer::FlightStart(), it won't be repopulated when the
mission restarts this way.

(cherry picked from commit 98dda84e92d283af37f5b6a465374047cf7dc122)