Use the BracketVR code to render a CubeMap.
I got a new superpower.

The bracket code can be used to render a big bracket, say, one that covers the
whole in-game screen.

Then the bracket code can be modified to remove the billboard transform, that
way, the bracket is always rendered in viewspace. Since the bracket is the size
of the whole screen, that means that the back-projected VR bracket covers the
whole VR FOV. We now have a large distant canvas that always faces the camera.

Then, just for giggles, let's render a fixed orientation in space on this large
canvas. The fixed orientation is computed by the Heading and Headset view
matrices. To display the fixed orientation, we compute a view vector using the
Big Canvas 3D position and take the dot product with the fixed orientation in
space. The result is a disk that stays put in place wrt the stellar background.
We have effectively created a way to render cubemaps!

Now, we can use this to replace the in-game's cubemap and use real environment
map. We should be able to do this for exterior views -- maybe even fix the
hangar background.

There's lots of work to do, but the concept works.