Align the brackets with the cockpit's axes.
This keeps the brackets stable and they don't roll anymore when looking around.
They are aligned with the cockpit, so that also mimics the way the work in
regular pancake mode.

I'm going to keep this version.

Next, I need to apply this fix when the external camera is active.