Capture the _constants at the beginning of every frame.
This works for the TextureCube because the constants we care about are just
the projection constants (the WorldView transform is set to identity). Using
the _frameConstants we can render the TextureCube even when there's no
Cockpit and no Exterior meshes (like in the hangar, in certain views).

This should also work for rendering other stuff that appears on the background.