Add replace_backdrops setting, BackdropId to GroupId-ImageId map.
If replace_backdrops is enabled, then we'll replace the backdrops using our own
engine (WIP). I also added an std::map that maps Backdrop Id's, as specified
in Yogeme to GroupId's (the ImageId is the same as the shadow index in the
backdrop). This will help us parse the XWA backrops and associate them with
images.

There's also code here to tag images that are used as backdrops, but I may
remove that later.