Add RenderSkyCylinder stub.
This code uses a modified copy of CacheBracketsVR()/RenderVRBrackets() to render
one side of a custom texture cube placed at infinity (65km). The plane is the
same bracket used to render the previous texture cube, but this plane rotates
according to the ship's heading and view. Works with the external view too, but
it won't work (yet) in VR mode.

I'll be using this code to render the XWA Sky Cylinder on my own to fix the
"swimming background" bug people have reported in VR when the external camera
is activated. Plus I may also use this to fix other artifacts that happen when
planets are too big.