Add TagXWABackdrops(). Backdrops are tagged separately.
Will likely need to use another std::map instead of just using counters to keep
track of the tagging. This is needed because the caps are not always visible
when first exiting the hangar. If we don't tag them, they won't show in the
SkyCylinder, so we need to keep checking until all cylinder sides have been
tagged.