SteamVR fix: Render the SkyBox after UpdateViewMatrix() has been called.
We need to render the default starfield in Direct3DDevice::BeginScene(), but we
have to do it after we update the view matrix, otherwise the camera will be all
wrong. We already call UpdateViewMatrix() in Direct3DDevice::Execute(), but
sometimes it doesn't get called at all (for instance, if there's no backdrops in
the current region). So we need to move the call to UpdateViewMatrix() into
BeginScene() and then we can call RenderSkyBox(). Plus, it makes more sense that
way.