Hardware-generated Mip-Map Levels for the Bloom buffer.
Looks like we can have the HW generate the mip-map levels for the Bloom input
buffer and then we can implement bloom using only 2 passes as implemented here:

https://www.shadertoy.com/view/cty3R3

This commit generates the mip-map levels and displays them using the LevelsPS
shader. A lot more work is needed, but the mip-maps work and we can sample any
level we want using SampleLevel in the shader.