Optimize the speed effect (WIP).
m0rgg raised a very good point a few days ago: we probably don't need to do a
CopyResource() when doing rendering the speed effect if we can render the speed
particles directly on the output buffer. The CopyResource() is an expensive
operation that should be avoided if possible.

The answer is yes: we can render the speed particles directly on the output
buffer but we need to provide the depthBuffer, post-proc UV coords and the Z
coord of the particles. We need this information to do a depth-test and reject
the particles that should be obscured by the cockpit.

We also need to use a post-proc viewport (i.e. a viewport that covers the whole
screen), and we need to provide the limits of the 3D viewport in post-proc UV
screen coords (sometimes they're not the same as there can be black bars on the
sides or the top/bottom). We need this to reject particles that would otherwise
be rendered on the black bars around the screen.

We also remove the second step when rendering this effect: we no longer need to
compose the particles, cockpit and 3D. So that's also great.

So, in a few words: yes, m0rgg was right. This is a nice optimization (thanks
for the idea m0rgg!) and this is a mixture of post-proc and forward-rendering
techniques which I had never attempted before. Fun!

WIP: I need to check different aspect ratios, full screen and VR.