Add TargetBoxHook() to detect current target bracket and subcomponent bracket.
DrawTargetBoxForPlayerTarget() calls TargetBox() in a number of places for both
target brackets and subcomponent brackets. TargetBox() calls
DrawTargetBoxInFlight() which was already hooked by Jeremy in
DrawBracketInFlightHook(). Thus by hooking the calls to TargetBox(), we can
inspect the target and the subcomponent indices and set global variables that
can later be used inside DrawBracketInFlightHook() to enhance the current target
bracket and the subcomponent bracket.

There are more places where TargetBox() is called that may need to be hooked;
but this appears to do the job for now.