Use RenderVRHUD() as a template to render VR enhanced HUD brackets.
RenderVRHUD() is currently not used, but when enabled, it will render a quad
right in front of the reticle. So, what we're doing here is we're rotating that
quad to align it with the current target bracket. Then we apply an additional
rotation to keep the up direction. Finally, another rotation is used to displace
the quad from the center of the target bracket so that we can display text there
later. Doing it this way avoids the apparent contraction when the brackets reach
the edge of the visual field. Still needs a bit of work, but this will work for
VR.