Add D3dRendererTexturesHook
Based on Jeremy's new hook, commit 7f3d736, see:

https://github.com/JeremyAnsel/xwa_ddraw_d3d11/commit/7f3d7363355586933ef87ab04692fb4587a526b8#diff-4125dc8693cd0f9dacb1608a7df55932cead35b0366292304d4da9c63d45382d

The diff was manually applied since the code has drifted significantly between
our branches.

Main difference in this implementation is that the EffectsRenderer.cpp also had
to be updated (see EffectsRenderer::UpdateTextures())

This is still WIP, the game will start but the effects are broken and a new
tagging system has to be implemented.