Fix transparency issue with the D3dRendererTexturesHook
When the D3dRendererTexturesHook is enabled, scene->TextureAlphaMask is always 0
This causes the transparent calls to be rendered solid (instead of using
transparency and deferred rendering). To fix this, we now need to check the
lastTextureSelected for transparency.

Also added the trailing ",0" to texture names -- I recall using that in
TagTextures() even if it's probably not used anymore.