Use RenderSkirmishOPT() to render the ship in the skirmish selection screen.
We need to add a new path to render just this case because we use array textures
in VR mode and _offscreenBufferHd is not an array texture. This fix works, but
it's incomplete: the aspect ratio is wrong and it's not using a Depth Stencil
(because we need to create one just for this path).