Enable backdrop overrides
Since I'm hard-coding the list of disabled backdrops for XWAU, there's a chance
that some of those backdrops should, in fact, be displayed later (or in TFTC).
So, I'm adding a new setting called "EnabledBackdrops" that will force specific
backdrops to be displayed even if they have been disabled before. Use it like
this:

	EnabledBackdrops = 6401-0