Enable "EnabledBackdrops = ALL" -- all backdrops are rendered!
To be fair, I'm not sure why people requested this feature, but it doesn't do
any harm, so here we go. In this case, adding:

	EnabledBackdrops = ALL

to the current mission's .ini file, will force ALL backdrops to be rendered when
CubeMaps are enabled. This means that the CubeMap itself won't be visible unless
the mission has been edited to remove stellar backdrops.