Fix the hangar and external camera CubeMap rotation/orientation.
Jeremy's code, properly applied this time. I had to capture the transform matrix
right at the beginning of the frame. Then we need to apply the proper transforms
to flip the axes and make it work. I think this was worth it since the hangar
rotation is now properly applied _and_ the external camera cubemaps also work
well now.

This code is enabled by default, to revert to the previous code, just add:

	use_default_skybox_orientation = 1

to SSAO.cfg