Fix CubeMap blending artifacts when transparent backdrops are present.
If a Sun is displayed over a backdrop with transparency, the blending won't
work. Also, blending transparent backdrops with CubeMaps shows some artifacts.
To fix this, I've enabled Jeremy's hook that renders CubeMaps before backdrops,
but I made it optional so that I don't break XWAU's cubemaps inadvertently. To
enable this fix, do the following:

1. Add: "RenderMode=Golden" in the affected mission's .ini file. This applies
   the fix only to the specific mission.
2. Add: "skybox_default_render_mode=Golden" to SSAO.cfg. This applies the fix
   to all missions

The fix is disabled by default so that I don't break anything by mistake.